Сделано рефакторинг кода. Реализовать активацию по коду
