+++ templates/account/email/email_confirmation_signup_message.html
@@ -7,7 +7,7 @@
 </head>
 <body>
     {% load account %} <!-- Загружаем тег который позволит получить нам ссылку на подтверждение -->
-    <h1>Спасибо за регистрацию {{ user.username }}!</h1>
+    <h1>Спасибо за регистрацию {{ user.first_name }}!</h1>
     <h3>Пожалуйста, следуйте <a href="{{ activate_url }}">по ссылке</a> чтобы активировать вашу учетную запись</h3>
      <!-- Всставляем ссылку на активацию -->
 </body>
